services/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-httpx>=0.42b0
opentelemetry-instrumentation-requests>=0.42b0
opentelemetry-exporter-otlp-proto-http>=1.21.0
prometheus-client>=0.19.0
httpx>=0.25.0
pydantic>=2.5.0
python-json-logger>=2.0.0


